swap bootScene and preloaderScene code
